Update with lists and edit views 06-09-2022

+++ Proyecto-Ciclo3.App/ProyectoCiclo3.App.Dominio/Entidades/Rutas.cs
@@ -4,6 +4,6 @@ namespace ProyectoCiclo3.App.Dominio{
         public int id { get; set; }
         public string origen { get; set; }
         public string destino { get; set; }
-        public int tiempo_estimado { get; set; }
+        public double tiempo_estimado { get; set; }
     } 
 }
